refactor webhook invoke to a class
