+++ packages/nocodb/src/services/v3/views-v3.service.ts
@@ -1,4 +1,5 @@
 import { Injectable, Logger } from '@nestjs/common';
+import type { MetaService } from '~/meta/meta.service';
 import type { NcContext, NcRequest } from '~/interface/config';
 import type { IViewsV3Service } from '~/services/v3/views-v3.types';
 
@@ -10,6 +11,7 @@ export class ViewsV3Service implements IViewsV3Service {
   async getView(
     _context: NcContext,
     _param: { viewId: string; req: NcRequest },
+    _ncMeta?: MetaService,
   ) {
     return undefined;
   }
